+++ src/main/java/ru/netology/javaqa/javaqamvn/services/SalesAnalysisService.java
@@ -1,21 +1,21 @@
 package ru.netology.javaqa.javaqamvn.services;
 
 public class SalesAnalysisService {
-    public int calculateAverage(int[] array) {
+    public long calculateAverage(long[] array) {
         if (array.length == 0) {
             throw new IllegalArgumentException("Array is empty");
         }
-        int sum = 0;
-        for (int num : array) {
+        long sum = 0;
+        for (long num : array) {
             sum += num;
         }
         return sum / array.length;
     }
 
-    public int countMonthsAboveAverage(int[] array) {
-        int average = calculateAverage(array);
+    public long countMonthsAboveAverage(long[] array) {
+        long average = calculateAverage(array);
         int count = 0;
-        for (int num : array) {
+        for (long num : array) {
             if (num > average) {
                 count++;
             }
@@ -23,20 +23,20 @@ public class SalesAnalysisService {
         return count;
     }
 
-    public int averageSum(int[] array) {
+    public long averageSum(long[] array) {
         if (array.length == 0) {
             throw new IllegalArgumentException("Array is empty");
         }
-        int sum = 0;
-        for (int num : array) {
+        long sum = 0;
+        for (long num : array) {
             sum += num;
         }
-        return (int) sum / array.length;
+        return (long) sum / array.length;
     }
 
-    public int belowAverage(int[] arr, int average) {
-        int count = 0;
-        for (int num : arr) {
+    public long belowAverage(long[] arr, long average) {
+        long count = 0;
+        for (long num : arr) {
             if (num < average) {
                 count++;
             }
@@ -45,37 +45,37 @@ public class SalesAnalysisService {
 
     }
 
-    public int sumSales(int[] sales) {
-        int sumAll = 0;
+    public long sumSales(long[] sales) {
+        long sumAll = 0;
 
-        for (int num : sales) {
+        for (long num : sales) {
             sumAll += num;
         }
         return sumAll;
     }
 
-    public int maxSales(int[] sales) {
-        int maxMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
+    public long maxSales(long[] sales) {
+        long maxMonth = 0;
 
-        for (int i = 0; i < sales.length; i++) {
-            if (sales[i] >= sales[maxMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
-                maxMonth = i; // запомним его как минимальный
+        for (long i = 0; i < sales.length; i++) {
+            if (sales[(int) i] >= sales[(int) maxMonth]) {
+                maxMonth = i;
             }
         }
 
-        return maxMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
+        return maxMonth + 1;
     }
 
-    public int minSales(int[] sales) {
-        int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
+    public long minSales(long[] sales) {
+        long minMonth = 0;
 
-        for (int i = 0; i < sales.length; i++) {
-            if (sales[i] <= sales[minMonth]) { // значит, в рассматриваемом i-м месяце продаж меньше
-                minMonth = i; // запомним его как минимальный
+        for (long i = 0; i < sales.length; i++) {
+            if (sales[(int) i] <= sales[(int) minMonth]) {
+                minMonth = i;
             }
         }
 
-        return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
+        return minMonth + 1;
     }
